backend/requirements.txt
asgiref==3.8.1
Django==5.2.3
django-cors==0.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
sqlparse==0.5.3
typing_extensions==4.14.0
djangorestframework-simplejwt==5.5.0
